Ignore unfixable issue in unused feature of nested dep
`chownr` is a dependency of `tar-fs`. We don't use tar-fs's "own" feature, therefore we don't use chownr. Snyk flags a vulnerability in `chownr` that is inherent to any recusive chown or chmod or similar (see https://github.com/isaacs/chownr/issues/14#issuecomment-451543037 ), and is only a problem on an already-compomised system where an attacker can swap a directory for a symlink between it being checked and being read.